+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UIProjectMembersService.java
@@ -4,24 +4,30 @@ import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 
+import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIConstraintViolationException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
 
 import ca.corefacility.bioinformatics.irida.exceptions.ProjectWithoutOwnerException;
+import ca.corefacility.bioinformatics.irida.model.enums.ProjectMetadataRole;
 import ca.corefacility.bioinformatics.irida.model.enums.ProjectRole;
 import ca.corefacility.bioinformatics.irida.model.joins.Join;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectUserJoin;
 import ca.corefacility.bioinformatics.irida.model.project.Project;
 import ca.corefacility.bioinformatics.irida.model.user.User;
+import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.NewMemberRequest;
 import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIProjectWithoutOwnerException;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.TableRequest;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.TableResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.projects.dto.ProjectMemberTableModel;
-import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.NewMemberRequest;
+import ca.corefacility.bioinformatics.irida.ria.web.utilities.ExceptionUtilities;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;
+
+import org.springframework.transaction.TransactionSystemException;
 
 /**
  * Service class for the UI for handling project members actions.
@@ -33,7 +39,8 @@ public class UIProjectMembersService {
 	private final MessageSource messageSource;
 
 	@Autowired
-	public UIProjectMembersService(ProjectService projectService, UserService userService, MessageSource messageSource) {
+	public UIProjectMembersService(ProjectService projectService, UserService userService,
+			MessageSource messageSource) {
 		this.projectService = projectService;
 		this.userService = userService;
 		this.messageSource = messageSource;
@@ -50,13 +57,11 @@ public class UIProjectMembersService {
 		Project project = projectService.read(projectId);
 		Page<Join<Project, User>> usersForProject = userService.searchUsersForProject(project, tableRequest.getSearch(),
 				tableRequest.getCurrent(), tableRequest.getPageSize(), tableRequest.getSort());
-		List<ProjectMemberTableModel> members = usersForProject.get()
-				.map(join -> {
-					ProjectUserJoin projectUserJoin = (ProjectUserJoin) join;
-					return new ProjectMemberTableModel(join.getObject(), projectUserJoin.getProjectRole()
-							.toString(), projectUserJoin.getCreatedDate());
-				})
-				.collect(Collectors.toList());
+		List<ProjectMemberTableModel> members = usersForProject.get().map(join -> {
+			ProjectUserJoin projectUserJoin = (ProjectUserJoin) join;
+			return new ProjectMemberTableModel(join.getObject(), projectUserJoin.getProjectRole().toString(),
+					((ProjectUserJoin) join).getMetadataRole().toString(), projectUserJoin.getCreatedDate());
+		}).collect(Collectors.toList());
 		return new TableResponse<>(members, usersForProject.getTotalElements());
 	}
 
@@ -69,7 +74,8 @@ public class UIProjectMembersService {
 	 * @return Message to display to the user about the outcome of removing a user from the project.
 	 * @throws UIProjectWithoutOwnerException if removing the user will leave the project without a manager
 	 */
-	public String removeUserFromProject(Long projectId, Long userId, Locale locale) throws UIProjectWithoutOwnerException {
+	public String removeUserFromProject(Long projectId, Long userId, Locale locale)
+			throws UIProjectWithoutOwnerException {
 		Project project = projectService.read(projectId);
 		User user = userService.read(userId);
 		try {
@@ -89,23 +95,73 @@ public class UIProjectMembersService {
 	 *
 	 * @param projectId - identifier for the current project
 	 * @param userId    - identifier for the user to remove from the project
-	 * @param role      - to update the user to
+	 * @param role      - {@link ProjectRole}  to update the user to
 	 * @param locale    - of the currently logged in user
 	 * @return message to display to the user about the outcome of the change in role.
 	 * @throws UIProjectWithoutOwnerException if removing the user will leave the project without a manager
+	 * @throws UIConstraintViolationException thrown when updating the project role to owner and the metadata role is
+	 *                                        not set to the highest level
 	 */
-	public String updateUserRoleOnProject(Long projectId, Long userId, String role, Locale locale) throws UIProjectWithoutOwnerException {
+	public String updateUserRoleOnProject(Long projectId, Long userId, String role, Locale locale)
+			throws UIProjectWithoutOwnerException, UIConstraintViolationException {
 		Project project = projectService.read(projectId);
 		User user = userService.read(userId);
+		ProjectMetadataRole projectMetadataRole;
 		ProjectRole projectRole = ProjectRole.fromString(role);
 		String roleString = messageSource.getMessage("projectRole." + role, new Object[] {}, locale);
 
-		try {
-			projectService.updateUserProjectRole(project, user, projectRole);
-			return messageSource.getMessage("server.ProjectRoleSelect.success", new Object[] { user.getLabel(), roleString }, locale);
-		} catch (ProjectWithoutOwnerException e) {
-			throw new UIProjectWithoutOwnerException(messageSource.getMessage("server.ProjectRoleSelect.error", new Object[] { user.getLabel(), roleString }, locale));
+		/*
+		 If a user's project role is set to collaborator we drop the metadata restriction to the lowest
+		 level and have the project owner set it accordingly. If a role of owner is set then that user is
+		 given full metadata permissions
+		 */
+		if (projectRole.equals(ProjectRole.PROJECT_OWNER)) {
+			projectMetadataRole = ProjectMetadataRole.fromString("LEVEL_4");
+		} else {
+			projectMetadataRole = ProjectMetadataRole.fromString("LEVEL_1");
 		}
+
+		try {
+			projectService.updateUserProjectRole(project, user, projectRole, projectMetadataRole);
+
+			return messageSource.getMessage("server.update.projectRole.success",
+					new Object[] { user.getLabel(), roleString }, locale);
+		} catch (ProjectWithoutOwnerException e) {
+			throw new UIProjectWithoutOwnerException(messageSource.getMessage("server.ProjectRoleSelect.error",
+					new Object[] { user.getLabel(), roleString }, locale));
+		} catch (TransactionSystemException e) {
+			ExceptionUtilities.throwConstraintViolationException(e, locale, messageSource);
+		}
+		return null;
+	}
+
+	/**
+	 * Update a users metadata role on a project
+	 *
+	 * @param projectId    - identifier for the current project
+	 * @param userId       - identifier for the user to remove from the project
+	 * @param metadataRole - {@link ProjectMetadataRole}  to update the user to
+	 * @param locale       - of the currently logged in user
+	 * @return message to display to the user about the outcome of the change in role.
+	 * @throws UIConstraintViolationException thrown when updating the project role to owner and the metadata role is
+	 *
+	 */
+	public String updateUserMetadataRoleOnProject(Long projectId, Long userId, String metadataRole, Locale locale)
+			throws UIConstraintViolationException {
+		Project project = projectService.read(projectId);
+		User user = userService.read(userId);
+		ProjectMetadataRole projectMetadataRole = ProjectMetadataRole.fromString(metadataRole);
+		String roleString = messageSource.getMessage("metadataRole." + metadataRole, new Object[] {}, locale);
+
+		try {
+			projectService.updateUserProjectMetadataRole(project, user, projectMetadataRole);
+			return messageSource.getMessage("server.update.metadataRole.success",
+					new Object[] { user.getLabel(), roleString }, locale);
+		} catch (TransactionSystemException e) {
+			ExceptionUtilities.throwConstraintViolationException(e, locale, messageSource);
+		}
+		return null;
+
 	}
 
 	/**
@@ -131,8 +187,16 @@ public class UIProjectMembersService {
 	public String addMemberToProject(Long projectId, NewMemberRequest request, Locale locale) {
 		Project project = projectService.read(projectId);
 		User user = userService.read(request.getId());
-		ProjectRole role = ProjectRole.fromString(request.getRole());
-		projectService.addUserToProject(project, user, role);
-		return messageSource.getMessage("project.members.add.success", new Object[] { user.getLabel(), project.getLabel() }, locale);
+		ProjectRole role = ProjectRole.fromString(request.getProjectRole());
+		ProjectMetadataRole metadataRole;
+
+		if (role.equals(ProjectRole.PROJECT_OWNER)) {
+			metadataRole = ProjectMetadataRole.LEVEL_4;
+		} else {
+			metadataRole = ProjectMetadataRole.fromString(request.getMetadataRole());
+		}
+		projectService.addUserToProject(project, user, role, metadataRole);
+		return messageSource.getMessage("project.members.add.success",
+				new Object[] { user.getLabel(), project.getLabel() }, locale);
 	}
 }