+++ src/main/java/ca/corefacility/bioinformatics/irida/security/permissions/IridaPermissionEvaluator.java
@@ -26,14 +26,14 @@ public class IridaPermissionEvaluator implements PermissionEvaluator {
 
 	private static final Logger logger = LoggerFactory.getLogger(IridaPermissionEvaluator.class);
 
-	private Collection<BasePermission<?,?>> permissions;
-	private Map<String, BasePermission<?,?>> namedPermissionMap;
+	private Collection<BasePermission<?>> permissions;
+	private Map<String, BasePermission<?>> namedPermissionMap;
 
-	public IridaPermissionEvaluator(BasePermission<?,?>... permissions) {
+	public IridaPermissionEvaluator(RepositoryBackedPermission<?,?>... permissions) {
 		this(Arrays.asList(permissions));
 	}
 
-	public IridaPermissionEvaluator(Collection<BasePermission<?,?>> permissions) {
+	public IridaPermissionEvaluator(Collection<BasePermission<?>> permissions) {
 		this.permissions = permissions;
 		this.namedPermissionMap = new HashMap<>();
 	}
@@ -43,7 +43,7 @@ public class IridaPermissionEvaluator implements PermissionEvaluator {
 	 */
 	@PostConstruct
 	public void init() {
-		for (BasePermission<?,?> p : permissions) {
+		for (BasePermission<?> p : permissions) {
 			logger.trace("Registering permission [" + p.getPermissionProvided() + "] with class ["
 					+ p.getClass().getName() + "]");
 			namedPermissionMap.put(p.getPermissionProvided(), p);
@@ -60,7 +60,7 @@ public class IridaPermissionEvaluator implements PermissionEvaluator {
 					+ "] is not registered with " + getClass().getName() + ".");
 		}
 
-		BasePermission<?,?> permissionEvaluator = namedPermissionMap.get(permission.toString());
+		BasePermission<?> permissionEvaluator = namedPermissionMap.get(permission.toString());
 		boolean allowed = permissionEvaluator.isAllowed(authentication, targetDomainObject);
 
 		logger.trace("Permission request for access to [" + targetDomainObject + "] with permission [" + permission