+++ src/main/webapp/resources/js/pages/projects/share/store.js
@@ -1,4 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
+import { fieldsApi } from "../../../apis/metadata/field";
 import { projectsApi } from "../../../apis/projects/projects";
 import { samplesApi } from "../../../apis/projects/samples";
 import shareReducer from "./shareSlice";
@@ -8,10 +9,12 @@ export default configureStore({
     shareReducer,
     [projectsApi.reducerPath]: projectsApi.reducer,
     [samplesApi.reducerPath]: samplesApi.reducer,
+    [fieldsApi.reducerPath]: fieldsApi.reducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat(
       projectsApi.middleware,
-      samplesApi.middleware
+      samplesApi.middleware,
+      fieldsApi.middleware
     ),
 });